pyproject.toml
[project]
name = "gsc-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.179.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "oauth2client>=4.1.3",
    "pandas>=2.3.2",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
]
